Proxy/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
WTForms==3.1.1
SQLAlchemy
Jinja2==3.1.2
Werkzeug==3.0.1
requests==2.31.0
openapi-spec-validator>=0.5.0
prance>=0.21.8
faker
phonenumbers
email_validator
pyap
cryptography
python-dateutil